+++ CHANGELOG.md
@@ -10,8 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - ReleaseDate
 
 - `dfxvm --list` now supports listing the remote available dfx versions.
-  - `--remote`: Specifies to list the available remote versions.
-  - `--number`: Specifies the number of remote versions to be listed in reverse chronological order, with default value `10`.
+  - `--available`: Specifies to list the remote available versions.
+  - `--limit`: Specifies the number of remote versions to be listed in reverse chronological order, with default value `10`.
 
 ## [1.0.0] - 2024-02-20